backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy>=2.0
pydantic>=2.0
python-multipart
python-jose[cryptography]
python-dotenv
pymupdf
sentence-transformers==2.2.2
weaviate-client>=4.0.0
openai>=1.0.0
huggingface_hub==0.10.1